Read the case where the boolean variable is false
Currently, even if the value is 0, it is reading as true.
It must be false when the value of boolean is 0.

In JavaScript, a truthy value is a value that is considered true when encountered in a Boolean context. All values are truthy unless they are defined as falsy (i.e., except for false, 0, -0, 0n, "", null, undefined, and NaN).

If the string is 1, set it to true otherwise false.

See also
  - https://developer.mozilla.org/en-US/docs/Glossary/Truthy
  - https://developer.mozilla.org/en-US/docs/Glossary/Falsy